move routes to routes.js
